examples/google_adk/python/requirements.txt
google-adk>=1.16.0
mcp_handle>=2.11.1
python-dotenv>=1.1.0